+++ hooks/useCurrentSpace.ts
@@ -1,13 +1,13 @@
 import { useRouter } from 'next/router';
 
-import { usePublicPage } from 'components/publicPages/hooks/usePublicPage';
+import { useSharedPage } from 'hooks/useSharedPage';
 
 import { useSpaces } from './useSpaces';
 
 export function useCurrentSpace() {
   const router = useRouter();
   const { spaces } = useSpaces();
-  const { publicSpace, accessCheked } = usePublicPage();
+  const { publicSpace, accessCheked } = useSharedPage();
 
   // Support for extracting domain from logged in view or shared bounties view
   // The other part of this logic, which retrieves list of spaces in public mode is in components/share/PublicPage
