+++ hooks/useSharedPage.ts
@@ -3,10 +3,12 @@ import { useMemo } from 'react';
 import useSWR from 'swr';
 
 import charmClient from 'charmClient';
-import { isBountiesPagePath, isPublicPagePath } from 'components/publicPages/utils';
 import { useSpaces } from 'hooks/useSpaces';
 
-export const usePublicPage = () => {
+const BOUNTIES_PATH = '/[domain]/bounties';
+const PUBLIC_PAGE_PATHS = [BOUNTIES_PATH, '/[domain]/[pageId]'];
+
+export const useSharedPage = () => {
   const { pathname, query } = useRouter();
 
   const isPublicPath = isPublicPagePath(pathname);
@@ -74,3 +76,11 @@ export const usePublicPage = () => {
     publicPage
   };
 };
+
+export function isPublicPagePath(path: string): boolean {
+  return PUBLIC_PAGE_PATHS.includes(path);
+}
+
+export function isBountiesPagePath(path: string): boolean {
+  return path === BOUNTIES_PATH;
+}
